Move numpy encoders to `dwave.cloud.utils.coders`
